allow receiving query parameters

+++ src/database/migrations/20220525223829_init/migration.sql
@@ -37,7 +37,7 @@ CREATE TABLE "recipes" (
     "steps" TEXT NOT NULL,
     "image" TEXT,
     "decoration" TEXT,
-    "alcohol_content" DECIMAL(65,30),
+    "alcohol_content" DOUBLE PRECISION,
     "original_url" TEXT,
     "hardness" INTEGER,
     "sweetness" INTEGER,
